+++ src/test/ExSln2/c99/gen/hal_LedArray.c
@@ -5,13 +5,23 @@
 
 
 
-void hal_LedArray_ctor(hal_LedArray * self, hal_Led * * leds)
+void hal_LedArray_ctor(hal_LedArray * self, hal_Led * * leds, uint8_t leds_length)
 {
     memset(self, 0, sizeof(*self));
-    self->leds = leds;
+    self->_leds = leds;
+    self->_leds_length = leds_length;
 }
 
 hal_Led * hal_LedArray_getLed(hal_LedArray * self, uint8_t index)
 {
-    return self->leds[index];
+    return self->_leds[index];
+}
+
+hal_Led * hal_LedArray_maybe_get_led(hal_LedArray * self, uint8_t index)
+{
+    if (index < self->_leds_length)
+    {
+        return self->_leds[index];
+    }
+    return NULL;
 }